+++ src/main/java/tetris/Controller.java
@@ -10,7 +10,6 @@ import tetris.tetromino.TetrominoFactory;
 import javafx.application.Platform;
 import javafx.scene.control.Button;
 import javafx.scene.input.KeyEvent;
-import javafx.scene.paint.Color;
 
 import highscore.GameEntry;
 import highscore.ScoreBoard;
@@ -25,8 +24,6 @@ public class Controller {
     private Grid grid;
     private AbstractTetromino tetromino;
     private AbstractTetromino tetromino2;
-    private int leftOffSet;
-    private int bottomOffSet;
     private boolean gameOver = false;
     private Tick timer = new Tick(event -> {
         Platform.runLater(() -> Action.SOFT_DROP.attempt(tetromino, grid));
@@ -187,85 +184,6 @@ public class Controller {
         timer.unpause();
         ui.resetFocus();
         Logger.log(this, Logger.LogType.INFO, "game restarted");
-    }
-
-    /**
-     * drawGrid draws the entire gameboard. As tetrominos reach their final
-     * place, they are registered on the grid to be drawn by this function.
-     */
-    private void drawGrid() {
-        for (int x = 0; x < settings.boardWidth(); x++) {
-            for (int y = 0; y < settings.boardHeight(); y++) {
-                drawRectangle(grid.get(x, y), new Coordinate(x, y));
-            }
-        }
-    }
-
-    /**
-     * drawTetromino employs the structure of a tetromino to draw it on a
-     * gameboard.
-     */
-    private void drawTetromino() {
-        for (int i = 0; i < 4; i++) {
-            drawRectangle(tetromino.getColor(), tetromino.get(i));
-        }
-    }
-
-    private void drawTetrominoPreview() {
-        for (int i = 0; i < 4; i++) {
-            drawRectanglePreview(tetromino2.getColor(), tetromino2.get(i));
-        }
-    }
-
-    /**
-     * drawRectangle draws one cube on the game grid.
-     * 
-     * @param color
-     *            specifies the color pair to draw in (color pairs provided by
-     *            setColor)
-     * @param coordinate
-     *            the cube in the grid that is to be drawn.
-     */
-    private void drawRectangle(int color, Coordinate coordinate) {
-        if (color > 0) {
-            settings.getBoard().setFill(settings.getColor(color));
-            settings.getBoard().fillRoundRect(coordinate.getX() * settings.blockSize(),
-                (settings.boardHeight() - 1 - coordinate.getY()) * settings.blockSize(),
-                settings.blockSize(), settings.blockSize(), settings.corner(), settings.corner());
-        }
-    }
-
-    /**
-     * drawRectanglePreview draws one cube on the preview grid.
-     *
-     * @param color
-     *            specifies the color pair to draw in (color pairs provided by
-     *            setColor)
-     * @param coordinate
-     *            the cube in the grid that is to be drawn.
-     */
-    private void drawRectanglePreview(int color, Coordinate coordinate) {
-        if (color > 0) {
-            settings.getPreview().setFill(settings.getColor(color));
-            settings.getPreview().fillRoundRect(
-                coordinate.getX() * settings.blockSize() + this.leftOffSet,
-                (5 - 1 - coordinate.getY()) * settings.blockSize() - this.bottomOffSet,
-                settings.blockSize(), settings.blockSize(), settings.corner(), settings.corner());
-        }
-    }
-
-    /**
-     * clearBoard erases the current board so it can be redrawn.
-     */
-    private void clearBoard() {
-        settings.getBoard().setFill(Color.BLACK);
-        settings.getBoard().fillRect(0, 0, settings.boardWidth() * settings.blockSize(),
-            settings.boardHeight() * settings.blockSize());
-    }
-
-    private void clearPreview() {
-        settings.getPreview().setFill(Color.BLACK);
-        settings.getPreview().fillRect(0, 0, 6 * settings.blockSize(), 5 * settings.blockSize());
     }
 
     public ScoreBoard getScoreBoard() {
