+++ src/main/java/tetris/scenes/GameScreen.java
@@ -9,6 +9,7 @@ import javafx.scene.Group;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.control.Button;
 import javafx.scene.layout.GridPane;
+import javafx.scene.layout.HBox;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 
@@ -40,26 +41,20 @@ public class GameScreen extends Group implements IScreen {
         this.settings = settings;
         VBox rightPane = setUpRightPaneGameScreen();
         Pane leftPane = setUpLeftPaneGameScreen();
-        GridPane rootGameScreen = setUpRootPaneGameScreen(leftPane, rightPane);
+        HBox rootGameScreen = setUpRootPaneGameScreen(leftPane, rightPane);
         getChildren().add(rootGameScreen);
         rootGameScreen.setStyle("-fx-background-color: grey");
     }
 
     private Pane setUpLeftPaneGameScreen() {
-        boardWidthPixel = settings.blockSize() * settings.boardWidth();
-        boardHeightPixel = settings.blockSize() * settings.boardHeight();
-        canvas = new Canvas(boardWidthPixel, boardHeightPixel);
-        settings.setBoard(canvas.getGraphicsContext2D());
+        canvas = new GridCanvas(settings);
         Pane leftPane = new Pane();
         leftPane.getChildren().add(canvas);
         return leftPane;
     }
 
-    private GridPane setUpRootPaneGameScreen(Pane leftPane, VBox rightPane) {
-        GridPane rootGameScreen = new GridPane();
-        rootGameScreen.setHgap(10);
-        GridPane.setConstraints(leftPane, 0, 0);
-        GridPane.setConstraints(rightPane, 1, 0);
+    private HBox setUpRootPaneGameScreen(Pane leftPane, VBox rightPane) {
+        HBox rootGameScreen = new HBox(10);
         rootGameScreen.getChildren().addAll(leftPane, rightPane);
         return rootGameScreen;
     }
